feat(SafeProposalDappDetailModal): remove ConfirmTransactionModal; bring user to ProposalBuilder directly
